+++ NEWS.md
@@ -1,3 +1,14 @@
+# qspray 2.0.0
+
+- The implementation of the opposite polynomial was wrong (it did nothing).
+
+- The pretty form of a polynomial returned by `prettyQspray` has been improved: now there are space around '+' and '-'.
+
+- The division of a polynomial by a `bigq` number did not always work.
+
+- New function `groebner`, to compute a Gröbner basis.
+
+
 # qspray 1.1.1
 
 Fixed a small error in the C++ code.